move user helper function into util helper
